+++ setup.cfg
@@ -1,5 +1,5 @@
 [metadata]
-version = 0.1.4
+version = 0.1.5.dev5
 license_files = LICENSE
 
 [tool.black]
